FIXES NEEDED
Making playerMax limit X and Y separately makes it so that the speed can
go higher than the max.

ex:  x = 10, y = 0 gives a velocity of 10 (which is fine)
but x = 10, y = 10 gives a velocity of 14.14 (which is faster!)

It's not that big of a deal now but it affects the controls and could
prove to be a problem later on.